requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
asyncio-timeout>=4.0.0
